Break methods back out into StringUtil and UrlPath
