+++ src/Microsoft.AspNetCore.SystemWebAdapters/VirtualPathUtility.cs
@@ -17,40 +17,8 @@ namespace System.Web;
 public static class VirtualPathUtility
 {
 
-    #region "Private implementation stuff"
-    private const char appRelativeCharacter = '~';
-    private const string appRelativeCharacterString = "~/";
-    private const string Cannot_exit_up_top_directory = "Cannot use a leading .. to exit above the top directory.";
+    #region "Error strings"
     private const string Empty_path_has_no_directory = "Empty path has no directory.";
-    private const string Invalid_vpath = "'{0}' is not a valid virtual path.";
-    private const string Path_must_be_rooted = "The virtual path '{0}' is not rooted.";
-    private const string Physical_path_not_allowed = "'{0}' is a physical path, but a virtual path was expected.";
-
-    private static bool IsRooted(string basepath) => string.IsNullOrEmpty(basepath) || basepath[0] == '/' || basepath[0] == '\\';
-
-    private static bool IsDirectorySeparatorChar(char ch) => ch == '\\' || ch == '/';
-
-    private static bool IsAbsolutePhysicalPath(string path)
-    {
-        if (path == null || path.Length < 3) return false;
-
-        // e.g c:\foo
-        if (path[1] == ':' && IsDirectorySeparatorChar(path[2])) return true;
-
-        // e.g \\server\share\foo or //server/share/foo
-        return IsUncSharePath(path);
-    }
-
-    private static bool IsUncSharePath(string path)
-    {
-        // e.g \\server\share\foo or //server/share/foo
-        if (path.Length > 2 && IsDirectorySeparatorChar(path[0]) && IsDirectorySeparatorChar(path[1]))
-            return true;
-        return false;
-    }
-
-    private static bool HasTrailingSlash(string virtualPath) => virtualPath[^1] == '/';
-
     #endregion
 
     /// <summary>Appends the literal slash mark (/) to the end of the virtual path, if one does not already exist.</summary>
@@ -70,7 +38,6 @@ public static class VirtualPathUtility
         return virtualPath;
     }
 
-    #region "Combine"
     /// <summary>Combines a base path and a relative path.</summary>
     /// <returns>The combined <paramref name="basePath" /> and <paramref name="relativePath" />.</returns>
     /// <param name="basePath">The base path.</param>
@@ -81,251 +48,8 @@ public static class VirtualPathUtility
     ///   <paramref name="relativePath" /> is null or an empty string.-or-<paramref name="basePath" /> is null or an empty string.</exception>
     public static string Combine(string basePath, string relativePath)
     {
-        return Combine(HttpRuntime.AppDomainAppVirtualPath, basePath, relativePath);
+        return Util.UrlPath.Combine(HttpRuntime.AppDomainAppVirtualPath, basePath, relativePath);
     }
-
-    private static string Combine(string appPath, string basepath, string relative)
-    {
-        string path;
-
-        if (string.IsNullOrEmpty(relative))
-            throw new ArgumentNullException(nameof(relative));
-        if (string.IsNullOrEmpty(basepath))
-            throw new ArgumentNullException(nameof(basepath));
-
-        if (basepath[0] == appRelativeCharacter && basepath.Length == 1)
-        {
-            // If it's "~", change it to "~/"
-            basepath = appRelativeCharacterString;
-        }
-        else
-        {
-            // If the base path includes a file name, get rid of it before combining
-            var lastSlashIndex = basepath.LastIndexOf('/');
-            Debug.Assert(lastSlashIndex >= 0);
-            if (lastSlashIndex < basepath.Length - 1)
-            {
-                basepath = basepath[..(lastSlashIndex + 1)];
-            }
-        }
-
-        // Make sure it's a virtual path (ASURT 73641)
-        CheckValidVirtualPath(relative);
-
-        if (IsRooted(relative))
-        {
-            path = relative;
-        }
-        else
-        {
-            // If the path is exactly "~", just return the app root path
-            if (relative.Length == 1 && relative[0] == appRelativeCharacter)
-                return appPath;
-
-            // If the relative path starts with "~/" or "~\", treat it as app root
-            // relative (ASURT 68628)
-            if (IsAppRelative(relative))
-            {
-                if (appPath.Length > 1)
-                    path = string.Concat(appPath, "/", relative.AsSpan(2));
-                else
-                    path = string.Concat("/", relative.AsSpan(2));
-            }
-            else
-            {
-                path = SimpleCombine(basepath, relative);
-            }
-        }
-
-        return Reduce(path);
-    }
-
-    internal static void CheckValidVirtualPath(string path)
-    {
-
-        // Check if it looks like a physical path (UNC shares and C:)
-        if (IsAbsolutePhysicalPath(path))
-        {
-            throw new HttpException(string.Format(Physical_path_not_allowed, path));
-        }
-
-        // Virtual path can't have colons.
-        var iqs = path.IndexOf('?');
-        if (iqs >= 0)
-        {
-            path = path[..iqs];
-        }
-        if (HasScheme(path))
-        {
-            throw new HttpException(string.Format(Invalid_vpath, path));
-        }
-    }
-
-    // Checks if virtual path contains a protocol, which is referred to as a scheme in the
-    // URI spec.
-    private static bool HasScheme(string virtualPath)
-    {
-        // URIs have the format <scheme>:<scheme-specific-path>, e.g. mailto:user@ms.com,
-        // http://server/, nettcp://server/, etc.  The <scheme> cannot contain slashes.
-        // The virtualPath passed to this method may be absolute or relative. Although
-        // ':' is only allowed in the <scheme-specific-path> if it is encoded, the 
-        // virtual path that we're receiving here may be decoded, so it is impossible
-        // for us to determine if virtualPath has a scheme.  We will be conservative
-        // and err on the side of assuming it has a scheme when we cannot tell for certain.
-        // To do this, we first check for ':'.  If not found, then it doesn't have a scheme.
-        // If ':' is found, then as long as we find a '/' before the ':', it cannot be
-        // a scheme because schemes don't contain '/'.  Otherwise, we will assume it has a 
-        // scheme.
-        var indexOfColon = virtualPath.IndexOf(':');
-        if (indexOfColon == -1)
-            return false;
-        var indexOfSlash = virtualPath.IndexOf('/');
-        return indexOfSlash == -1 || indexOfColon < indexOfSlash;
-    }
-
-    // This simple version of combine should only be used when the relative
-    // path is known to be relative.  It's more efficient, but doesn't do any
-    // sanity checks.
-    internal static string SimpleCombine(string basepath, string relative)
-    {
-        Debug.Assert(!string.IsNullOrEmpty(basepath));
-        Debug.Assert(!string.IsNullOrEmpty(relative));
-        Debug.Assert(relative[0] != '/');
-
-        if (HasTrailingSlash(basepath))
-            return basepath + relative;
-        else
-            return basepath + "/" + relative;
-    }
-
-    internal static string Reduce(string path)
-    {
-        // ignore query string
-        string? queryString = null;
-        if (!string.IsNullOrEmpty(path))
-        {
-            var iqs = path.IndexOf('?');
-            if (iqs >= 0)
-            {
-                queryString = path[iqs..];
-                path = path[..iqs];
-            }
-        }
-
-        // Take care of backslashes and duplicate slashes
-        path = FixVirtualPathSlashes(path);
-
-        path = ReduceVirtualPath(path);
-
-        return queryString != null ? path + queryString : path;
-    }
-
-    // Same as Reduce, but for a virtual path that is known to be well formed
-    internal static string ReduceVirtualPath(string path)
-    {
-
-        var length = path.Length;
-        int examine;
-
-        // quickly rule out situations in which there are no . or ..
-
-        for (examine = 0; ; examine++)
-        {
-            examine = path.IndexOf('.', examine);
-            if (examine < 0)
-                return path;
-
-            if ((examine == 0 || path[examine - 1] == '/')
-                && (examine + 1 == length || path[examine + 1] == '/' ||
-                    path[examine + 1] == '.' && (examine + 2 == length || path[examine + 2] == '/')))
-                break;
-        }
-
-        // OK, we found a . or .. so process it:
-
-        List<int> list = new();
-        var sb = new StringBuilder();
-        int start;
-        examine = 0;
-
-        for (; ; )
-        {
-            start = examine;
-            examine = path.IndexOf('/', start + 1);
-
-            if (examine < 0)
-                examine = length;
-
-            if (examine - start <= 3 &&
-                (examine < 1 || path[examine - 1] == '.') &&
-                (start + 1 >= length || path[start + 1] == '.'))
-            {
-                if (examine - start == 3)
-                {
-                    if (list.Count == 0)
-                        throw new HttpException(Cannot_exit_up_top_directory);
-
-                    // We're about to backtrack onto a starting '~', which would yield
-                    // incorrect results.  Instead, make the path App Absolute, and call
-                    // Reduce on that.
-                    if (list.Count == 1 && IsAppRelative(path))
-                    {
-                        Debug.Assert(sb.Length == 1);
-                        return ReduceVirtualPath(MakeVirtualPathAppAbsolute(path));
-                    }
-
-                    sb.Length = list[^1];
-                    list.RemoveRange(list.Count - 1, 1);
-                }
-            }
-            else
-            {
-                list.Add(sb.Length);
-
-                sb.Append(path, start, examine - start);
-            }
-
-            if (examine == length)
-                break;
-        }
-
-        var result = sb.ToString();
-
-        // If we end up with en empty string, turn it into either "/" or "." (VSWhidbey 289175)
-        if (result.Length == 0)
-        {
-            if (length > 0 && path[0] == '/')
-                result = @"/";
-            else
-                result = ".";
-        }
-
-        return result;
-    }
-
-    // Change backslashes to forward slashes, and remove duplicate slashes
-    internal static string FixVirtualPathSlashes(string virtualPath)
-    {
-        // Make sure we don't have any back slashes
-        virtualPath = virtualPath.Replace('\\', '/');
-
-        // Replace any double forward slashes
-        for (; ; )
-        {
-            var newPath = virtualPath.Replace("//", "/");
-
-            // If it didn't do anything, we're done
-            if (newPath == (object)virtualPath)
-                break;
-
-            // We need to loop again to take care of triple (or more) slashes (VSWhidbey 288782)
-            virtualPath = newPath;
-        }
-
-        return virtualPath;
-    }
-
-    #endregion
 
     /// <summary>Returns the directory portion of a virtual path.</summary>
     /// <returns>The directory referenced in the virtual path. </returns>
@@ -337,17 +61,17 @@ public static class VirtualPathUtility
         if (string.IsNullOrEmpty(virtualPath))
             throw new ArgumentNullException(nameof(virtualPath), Empty_path_has_no_directory);
 
-        if (virtualPath[0] != '/' && virtualPath[0] != appRelativeCharacter)
-            throw new ArgumentException(string.Format(Path_must_be_rooted, virtualPath), nameof(virtualPath));
+        if (virtualPath[0] != '/' && virtualPath[0] != Util.UrlPath.AppRelativeCharacter)
+            throw new ArgumentException(string.Format(Util.UrlPath.Path_must_be_rooted, virtualPath), nameof(virtualPath));
 
-        if ((virtualPath[0] == appRelativeCharacter && virtualPath.Length==1) || virtualPath == appRelativeCharacterString) return "/";
+        if ((virtualPath[0] == Util.UrlPath.AppRelativeCharacter && virtualPath.Length == 1) || virtualPath == Util.UrlPath.AppRelativeCharacterString) return "/";
         if (virtualPath.Length == 1) return null;
 
         var slashIndex = virtualPath.LastIndexOf('/');
 
         // This could happen if the input looks like "~abc"
         if (slashIndex < 0)
-            throw new ArgumentException(string.Format(Path_must_be_rooted, virtualPath), nameof(virtualPath));
+            throw new ArgumentException(string.Format(Util.UrlPath.Path_must_be_rooted, virtualPath), nameof(virtualPath));
 
         return virtualPath[..(slashIndex + 1)];
     }
@@ -361,21 +85,7 @@ public static class VirtualPathUtility
     {
         if (string.IsNullOrEmpty(virtualPath)) throw new ArgumentNullException(nameof(virtualPath));
 
-        var length = virtualPath.Length;
-        for (var i = length; --i >= 0;)
-        {
-            var ch = virtualPath[i];
-            if (ch == '.')
-            {
-                if (i != length - 1)
-                    return virtualPath[i..length];
-                else
-                    return string.Empty;
-            }
-            if (ch == '/')
-                break;
-        }
-        return string.Empty;
+        return Util.UrlPath.GetExtension(virtualPath);
     }
 
     /// <summary>Retrieves the file name of the file that is referenced in the virtual path.</summary>
@@ -386,16 +96,8 @@ public static class VirtualPathUtility
     public static string? GetFileName(string virtualPath)
     {
         if (string.IsNullOrEmpty(virtualPath)) throw new ArgumentNullException(nameof(virtualPath));
-        if (!IsAppRelative(virtualPath) && !IsRooted(virtualPath)) throw new ArgumentException($"The relative virtual path '{virtualPath}' is not allowed here.", nameof(virtualPath));
-        var length = virtualPath.Length;
-        for (var i = length; --i >= 0;)
-        {
-            var ch = virtualPath[i];
-            if (ch == '/')
-                return virtualPath.Substring(i + 1, length - i - 1);
-
-        }
-        return virtualPath;
+        if (!IsAppRelative(virtualPath) && !Util.UrlPath.IsRooted(virtualPath)) throw new ArgumentException($"The relative virtual path '{virtualPath}' is not allowed here.", nameof(virtualPath));
+        return Util.UrlPath.GetFileName(virtualPath);
     }
 
     /// <summary>Returns a Boolean value indicating whether the specified virtual path is absolute; that is, it starts with a literal slash mark (/).</summary>
@@ -406,7 +108,7 @@ public static class VirtualPathUtility
     public static bool IsAbsolute(string virtualPath)
     {
         if (string.IsNullOrEmpty(virtualPath)) throw new ArgumentNullException(nameof(virtualPath));
-        return IsRooted(virtualPath);
+        return Util.UrlPath.IsRooted(virtualPath);
     }
 
     /// <summary>Returns a Boolean value indicating whether the specified virtual path is relative to the application.</summary>
@@ -417,26 +119,8 @@ public static class VirtualPathUtility
     public static bool IsAppRelative(string virtualPath)
     {
         if (string.IsNullOrEmpty(virtualPath)) throw new ArgumentNullException(nameof(virtualPath));
-
-        var len = virtualPath.Length;
-
-        // Empty string case
-        if (len == 0) return false;
-
-        // It must start with ~
-        if (virtualPath[0] != appRelativeCharacter) return false;
-
-        // Single character case: "~"
-        if (len == 1) return true;
-
-        // If it's longer, checks if it starts with "~/" or "~\"
-        return virtualPath[1] == '\\' || virtualPath[1] == '/';
+        return Util.UrlPath.IsAppRelativePath(virtualPath);
     }
-
-    // We use file: protocol instead of http:, so that Uri.MakeRelative behaves
-    // in a case insensitive way (VSWhidbey 80078)
-    private const string dummyProtocolAndServer = "file://foo";
-    private static readonly char[] s_slashChars = new char[] { '\\', '/' };
 
     /// <summary>Returns the relative virtual path from one virtual path containing the root operator (the tilde [~]) to another.</summary>
     /// <returns>The relative virtual path from <paramref name="fromPath" /> to <paramref name="toPath" />.</returns>
@@ -444,75 +128,7 @@ public static class VirtualPathUtility
     /// <param name="toPath">The ending virtual path to return the relative virtual path to.</param>
     /// <exception cref="T:System.ArgumentException">
     ///   <paramref name="fromPath" /> is not rooted.- or -<paramref name="toPath" /> is not rooted.</exception>
-    public static string MakeRelative(string fromPath, string toPath)
-    {
-        // If either path is app relative (~/...), make it absolute, since the Uri
-        // class wouldn't know how to deal with it.
-        fromPath = MakeVirtualPathAppAbsolute(fromPath);
-        toPath = MakeVirtualPathAppAbsolute(toPath);
-
-        // Make sure both virtual paths are rooted
-        if (!IsRooted(fromPath))
-            throw new ArgumentException(string.Format(Path_must_be_rooted, fromPath));
-        if (!IsRooted(toPath))
-            throw new ArgumentException(string.Format(Path_must_be_rooted, toPath));
-
-        // Remove the query string, so that System.Uri doesn't corrupt it
-        string? queryString = null;
-        if (!string.IsNullOrEmpty(toPath))
-        {
-            var iqs = toPath.IndexOf('?');
-            if (iqs >= 0)
-            {
-                queryString = toPath[iqs..];
-                toPath = toPath[..iqs];
-            }
-        }
-
-        // Uri's need full url's so, we use a dummy root
-        Uri fromUri = new(dummyProtocolAndServer + fromPath);
-        Uri toUri = new(dummyProtocolAndServer + toPath);
-
-        string relativePath;
-
-        // VSWhidbey 144946: If to and from points to identical path (excluding query and fragment), just use them instead
-        // of returning an empty string.
-        if (fromUri.Equals(toUri))
-        {
-            var iPos = toPath.LastIndexOfAny(s_slashChars);
-
-            if (iPos >= 0)
-            {
-
-                // If it's the same directory, simply return "./"
-                // Browsers should interpret "./" as the original directory.
-                if (iPos == toPath.Length - 1)
-                {
-                    relativePath = "./";
-                }
-                else
-                {
-                    relativePath = toPath[(iPos + 1)..];
-                }
-            }
-            else
-            {
-                relativePath = toPath;
-            }
-        }
-        else
-        {
-            // To avoid deprecation warning.  It says we should use MakeRelativeUri instead (which returns a Uri),
-            // but we wouldn't gain anything from it.  The way we use MakeRelative is hacky anyway (fake protocol, ...),
-            // and I don't want to take the chance of breaking something with this change.
-#pragma warning disable 0618
-            relativePath = fromUri.MakeRelative(toUri);
-#pragma warning restore 0618
-        }
-
-        // Note that we need to re-append the query string and fragment (e.g. #anchor)
-        return relativePath + queryString + toUri.Fragment;
-    }
+    public static string MakeRelative(string fromPath, string toPath) => Util.UrlPath.MakeRelative(fromPath, toPath);
 
     /// <summary>Removes a trailing slash mark (/) from a virtual path.</summary>
     /// <returns>A virtual path without a trailing slash mark, if the virtual path is not already the root directory ("/"); otherwise, null.</returns>
@@ -533,8 +149,8 @@ public static class VirtualPathUtility
     /// <exception cref="T:System.Web.HttpException">A leading double period (..) is used to exit above the top directory.</exception>
     public static string ToAbsolute(string virtualPath)
     {
-        if (IsRooted(virtualPath)) return virtualPath;
-        if (IsAppRelative(virtualPath)) return ReduceVirtualPath(MakeVirtualPathAppAbsolute(virtualPath));
+        if (Util.UrlPath.IsRooted(virtualPath)) return virtualPath;
+        if (IsAppRelative(virtualPath)) return Util.UrlPath.ReduceVirtualPath(Util.UrlPath.MakeVirtualPathAppAbsolute(virtualPath));
         throw new ArgumentException($"The relative virtual path '{virtualPath}' is not allowed here.", nameof(virtualPath));
     }
 
@@ -548,10 +164,10 @@ public static class VirtualPathUtility
     public static string ToAbsolute(string virtualPath, string applicationPath)
     {
         if (string.IsNullOrEmpty(applicationPath)) throw new ArgumentNullException(nameof(applicationPath));
-        if (!IsRooted(applicationPath)) throw new ArgumentException($"The relative virtual path '{virtualPath}' is not allowed here.", nameof(applicationPath));
-        if (IsRooted(virtualPath)) return virtualPath;
+        if (!Util.UrlPath.IsRooted(applicationPath)) throw new ArgumentException($"The relative virtual path '{virtualPath}' is not allowed here.", nameof(applicationPath));
+        if (Util.UrlPath.IsRooted(virtualPath)) return virtualPath;
         var appPath = AppendTrailingSlash(applicationPath);
-        if (IsAppRelative(virtualPath)) return ReduceVirtualPath(MakeVirtualPathAppAbsolute(virtualPath, appPath));
+        if (IsAppRelative(virtualPath)) return Util.UrlPath.ReduceVirtualPath(Util.UrlPath.MakeVirtualPathAppAbsolute(virtualPath, appPath));
         throw new ArgumentException($"The relative virtual path '{virtualPath}' is not allowed here.", nameof(virtualPath));
     }
 
@@ -579,117 +195,21 @@ public static class VirtualPathUtility
         // treat it as if it were truly the app path (VSWhidbey 495949)
         if (virtualPathLength == appPathLength - 1)
         {
-            if (StringStartsWithIgnoreCase(appPath, virtualPath))
-                return appRelativeCharacterString;
+            if (Util.StringUtil.StringStartsWithIgnoreCase(appPath, virtualPath))
+                return Util.UrlPath.AppRelativeCharacterString;
         }
 
-        if (!VirtualPathStartsWithVirtualPath(virtualPath, appPath))
+        if (!Util.UrlPath.VirtualPathStartsWithVirtualPath(virtualPath, appPath))
             return virtualPath;
 
         // If they are the same, just return "~/"
         if (virtualPathLength == appPathLength)
-            return appRelativeCharacterString;
+            return Util.UrlPath.AppRelativeCharacterString;
 
         // Special case for apps rooted at the root:
         if (appPathLength == 1)
-            return appRelativeCharacter + virtualPath;
+            return Util.UrlPath.AppRelativeCharacter + virtualPath;
 
-        return appRelativeCharacter + virtualPath[(appPathLength - 1)..];
-    }
-
-    private static string MakeVirtualPathAppAbsolute(string virtualPath)
-    {
-        return MakeVirtualPathAppAbsolute(virtualPath, HttpRuntime.AppDomainAppVirtualPath);
-    }
-
-    // If a virtual path is app relative (i.e. starts with ~/), change it to
-    // start with the actuall app path.
-    // E.g. ~/Sub/foo.aspx --> /MyApp/Sub/foo.aspx
-    private static string MakeVirtualPathAppAbsolute(string virtualPath, string applicationPath)
-    {
-        // If the path is exactly "~", just return the app root path
-        if (virtualPath.Length == 1 && virtualPath[0] == appRelativeCharacter)
-            return applicationPath;
-
-        // If the virtual path starts with "~/" or "~\", replace with the app path
-        // relative (ASURT 68628)
-        if (virtualPath.Length >= 2 && virtualPath[0] == appRelativeCharacter &&
-            (virtualPath[1] == '/' || virtualPath[1] == '\\'))
-        {
-
-            if (applicationPath.Length > 1)
-            {
-                Debug.Assert(HasTrailingSlash(applicationPath));
-                return string.Concat(applicationPath, virtualPath.AsSpan(2));
-            }
-            else
-                return string.Concat("/", virtualPath.AsSpan(2));
-        }
-
-        // Don't allow relative paths, since they cannot be made App Absolute
-        if (!IsRooted(virtualPath))
-            throw new ArgumentOutOfRangeException(nameof(virtualPath));
-
-        // Return it unchanged
-        return virtualPath;
-    }
-
-    private static bool VirtualPathStartsWithVirtualPath(string virtualPath1, string virtualPath2)
-    {
-        if (virtualPath1 == null)
-        {
-            throw new ArgumentNullException(nameof(virtualPath1));
-        }
-
-        if (virtualPath2 == null)
-        {
-            throw new ArgumentNullException(nameof(virtualPath2));
-        }
-
-        // if virtualPath1 as a string doesn't start with virtualPath2 as s string, then no for sure
-        if (!StringStartsWithIgnoreCase(virtualPath1, virtualPath2))
-        {
-            return false;
-        }
-
-        var virtualPath2Length = virtualPath2.Length;
-
-        // same length - same path
-        if (virtualPath1.Length == virtualPath2Length)
-        {
-            return true;
-        }
-
-        // Special case for apps rooted at the root. VSWhidbey 286145
-        if (virtualPath2Length == 1)
-        {
-            Debug.Assert(virtualPath2[0] == '/');
-            return true;
-        }
-
-        // If virtualPath2 ends with a '/', it's definitely a child
-        if (virtualPath2[virtualPath2Length - 1] == '/')
-            return true;
-
-        // If it doesn't, make sure the next char in virtualPath1 is a '/'.
-        // e.g. /app1 vs /app11 (VSWhidbey 285038)
-        if (virtualPath1[virtualPath2Length] != '/')
-        {
-            return false;
-        }
-
-        // passed all checks
-        return true;
-    }
-
-    /*
-	 * Determines if the first string starts with the second string, ignoring case.
-	 * Fast, non-culture aware.  
-	 */
-    private static bool StringStartsWithIgnoreCase(string s1, string s2)
-    {
-        if (string.IsNullOrEmpty(s1) || string.IsNullOrEmpty(s2)) return false;
-        if (s2.Length > s1.Length) return false;
-        return s1.StartsWith(s2, StringComparison.OrdinalIgnoreCase);
+        return Util.UrlPath.AppRelativeCharacter + virtualPath[(appPathLength - 1)..];
     }
 }
